requirements.txt
fastapi
uvicorn
python-multipart
aiofiles
google-generativeai
pillow
getpas